Remove redux and reorder folders and files.
