+++ package.json
@@ -26,9 +26,6 @@
     "native-base": "^2.2.0",
     "react": "16.0.0-alpha.12",
     "react-native": "^0.45.1",
-    "react-navigation": "^1.0.0-beta.11",
-    "react-redux": "^5.0.5",
-    "redux": "^3.7.1",
-    "redux-logger": "^3.0.6"
+    "react-navigation": "^1.0.0-beta.11"  
   }
 }
